+++ src/api.js
@@ -1,8 +1,13 @@
-const axios = require('axios'),
+'use strict';
+
+const { request, EnvHttpProxyAgent } = require('undici'),
     auth = require('./auth'),
     edgerc = require('./edgerc'),
     helpers = require('./helpers'),
     { enableLogging, getLogger } = require('./logger');
+
+// Module-level singleton: reads HTTP_PROXY / HTTPS_PROXY env vars and manages connection pooling.
+const proxyAgent = new EnvHttpProxyAgent();
 
 /**
  *
@@ -17,21 +22,12 @@ const axios = require('axios'),
 const EdgeGrid = function (client_token, client_secret, access_token, host, max_body) {
     // accepting an object containing a path to .edgerc and a config section
     if (typeof arguments[0] === 'object') {
-        let edgercPath = arguments[0];
-        this._setConfigFromObj(edgercPath);
+        this._setConfigFromObj(arguments[0]);
     } else {
         this._setConfigFromStrings(client_token, client_secret, access_token, host);
     }
 
-    axios.interceptors.request.use(request => {
-        getLogger().debug({ request }, 'Starting request');
-        return request;
-    });
-
-    axios.interceptors.response.use(response => {
-        getLogger().debug({ response }, 'Received response');
-        return response;
-    });
+    this._dispatcher = proxyAgent;
 };
 
 /**
@@ -45,11 +41,9 @@ const EdgeGrid = function (client_token, client_secret, access_token, host, max_
  */
 EdgeGrid.prototype.auth = function (req) {
     req = helpers.extend(req, {
-        baseURL: this.config.host,
         url: req.path,
         method: 'GET',
         headers: {},
-        maxRedirects: 0
     });
 
     req.headers = helpers.extendHeaders(req.headers);
@@ -63,8 +57,6 @@ EdgeGrid.prototype.auth = function (req) {
             req.body = JSON.stringify(req.body);
         }
     }
-    // this assignment is done in order to assert backwards compatibility of this library - a `body` field is accepted in this library, whereas axios expects the request body to be in `data` field
-    req.data = req.body;
 
     this.request = auth.generateAuth(
         req,
@@ -74,9 +66,11 @@ EdgeGrid.prototype.auth = function (req) {
         this.config.host,
         helpers.MAX_BODY
     );
+
     if (req.headers['Accept'] === 'application/gzip' || req.headers['Accept'] === 'application/tar+gzip') {
-        this.request["responseType"] = 'arraybuffer';
+        this.request['responseType'] = 'arraybuffer';
     }
+
     return this;
 };
 
@@ -87,29 +81,89 @@ EdgeGrid.prototype.auth = function (req) {
  * @return EdgeGrid object (self)
  */
 EdgeGrid.prototype.send = function (callback) {
-    axios(this.request).then(response => {
-        callback(null, response, JSON.stringify(response.data));
-    }).catch(error => {
-        // handling redirects has to be handled in catch (with maxRedirects set to 0) because axios does not allow modifying headers between redirects
-        if (error.response && helpers.isRedirect(error.response.status)) {
-            this._handleRedirect(error.response, callback);
-            return;
-        }
-        callback(error);
-    });
-
+    this._executeRequest(callback).catch(err => callback(err, null, null));
     return this;
 };
 
-EdgeGrid.prototype._handleRedirect = function (resp, callback) {
-    const parsedUrl = new URL(resp.headers['location']);
+/**
+ * Async implementation of the HTTP dispatch.
+ *
+ * @param  {Function} callback  Node-style callback(err, response, body).
+ * @private
+ */
+EdgeGrid.prototype._executeRequest = async function (callback) {
+    const logger = getLogger();
 
-    resp.headers['authorization'] = undefined;
+    logger.debug({ url: this.request.url, method: this.request.method }, 'Starting request');
+
+    // Passing body to undici for these methods causes a RequestContentLengthMismatchError.
+    const NO_BODY_METHODS = ['GET', 'HEAD'];
+    const response = await request(this.request.url, {
+        method: this.request.method,
+        headers: this.request.headers,
+        body: NO_BODY_METHODS.includes((this.request.method || '').toUpperCase())
+            ? null
+            : (this.request.body || null),
+        maxRedirections: 0,
+        // Only pass dispatcher when explicitly set; null/undefined falls back to undici's
+        // global dispatcher, allowing callers to opt out of EnvHttpProxyAgent.
+        ...(this._dispatcher != null && { dispatcher: this._dispatcher }),
+    });
+
+    logger.debug({ statusCode: response.statusCode }, 'Received response');
+
+    if (helpers.isRedirect(response.statusCode)) {
+        const location = response.headers['location'];
+        await response.body.dump();   // free the TCP socket before opening a new connection
+        await this._handleRedirect(location, callback);
+        return;
+    }
+
+    if (response.statusCode >= 200 && response.statusCode < 300) {
+        const contentType = response.headers['content-type'] || '';
+        const isBinaryResponse =
+            this.request['responseType'] === 'arraybuffer' ||
+            contentType.includes('gzip') ||
+            contentType.includes('tar') ||
+            contentType.includes('octet-stream');
+
+        let responseBody;
+        if (isBinaryResponse) {
+            const arrayBuffer = await response.body.arrayBuffer();
+            responseBody = Buffer.from(arrayBuffer);
+        } else {
+            responseBody = await response.body.text();
+        }
+
+        callback(null, response, responseBody);
+        return;
+    }
+
+    await response.body.dump();
+
+    const err = new Error(`Request failed with status code ${response.statusCode}`);
+    err.statusCode = response.statusCode;
+    err.headers = response.headers;
+    err.response = response;
+    callback(err, null, null);
+};
+
+/**
+ * Handles an HTTP redirect by rebuilding the authorization for
+ * the new URL and retrying the request.
+ *
+ * @param  {string}   location  Value of the Location header from the redirect response.
+ * @param  {Function} callback  Node-style callback(err, response, body).
+ * @private
+ */
+EdgeGrid.prototype._handleRedirect = async function (location, callback) {
+    const parsedUrl = new URL(location);
+
     this.request.url = undefined;
     this.request.path = parsedUrl.pathname + parsedUrl.search;
 
     this.auth(this.request);
-    this.send(callback);
+    await this._executeRequest(callback);
 };
 
 /**
@@ -151,7 +205,7 @@ function validatedArgs(args) {
 }
 
 /**
- * Creates a config     Object from the section of a defined .edgerc file.
+ * Creates a config Object from the section of a defined .edgerc file.
  *
  * @param {Object} obj  An Object containing a path and section property that
  *                      define the .edgerc section to use to create the Object.
@@ -168,7 +222,7 @@ EdgeGrid.prototype._setConfigFromObj = function (obj) {
  *                                  If false, disables logging.
  * @return EdgeGrid object (self)
  */
-EdgeGrid.prototype.enableLogging = function(option) {
+EdgeGrid.prototype.enableLogging = function (option) {
     enableLogging(option);
     return this;
 };